requirements.txt
aiogram
google-auth
google-api-python-client
apscheduler
difflib
